refactor: Make dist/mcp.js directly executable
- Extract MCP server implementation to src/mcp-server.ts
- Make mcp.ts a simple entry point that starts the server
- Update CLI to import from src/mcp-server.ts
- Now dist/mcp.js can be executed directly with node

This ensures the DXT package works correctly as dist/mcp.js
is now a standalone executable file.
